requirements.txt
Flask==3.1.1
langchain==0.3.25
langchain_community==0.3.24
langchain_core==0.3.64
langchain_google_genai==2.1.5
langchain_text_splitters==0.3.8
python-dotenv==1.1.0
tqdm==4.66.5
Werkzeug==3.1.3
